+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Bootsy Changelog
 
+## master branch
+
+* Merged the [SimpleForm](https://github.com/plataformatec/simple_form) support into master.
+  You no longer need to use the [bootsy-simple_form](https://github.com/volmer/bootsy-simple_form)
+  gem in your project.
+
 ## 1.1.0
 
 * Design and interaction improvements. Bootsy is a lot more 'ajaxy' now, and loads images